% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedHeatmap.R
\name{annotatedHeatmap}
\alias{annotatedHeatmap}
\title{Annotated heatmap}
\usage{
annotatedHeatmap(
  x,
  cluster_IDs,
  col_pal = colorRampPalette(c("#146EB4", "white", "#FF9900"))(100),
  my_breaks = defineDataBreaks(x, col_pal, mid_point = 0),
  main = "gen_dataset",
  ...
)
}
\arguments{
\item{x}{A matrix.}

\item{cluster_IDs}{A vector of the labels associated with the rows of x.}

\item{col_pal}{The colour palette to use in the heatmap.}

\item{my_breaks}{The breaks to use within the heatmap.}

\item{...}{Other arguments to pass to ``pheatmap``.}
}
\description{
Generate an annotated heatmap for output of
``generateGaussianDataset`` or ``generateComplexDataset`` using the
``pheatmap`` function.
}
