% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arandi4Apply.R, R/arandi_matrices.R
\name{arandi4Apply}
\alias{arandi4Apply}
\title{Arandi 4 apply}
\usage{
arandi4Apply(my_vec, break_col, end_col)

arandi4Apply(my_vec, break_col, end_col)
}
\description{
Arandi 4 apply

Arandi 4 apply
}
