% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrelationPlot.R
\name{autoCorrelationPlot}
\alias{autoCorrelationPlot}
\title{Auto-correlation plot}
\usage{
autoCorrelationPlot(x, plot_title, lag_max = NULL, facet = FALSE)
}
\arguments{
\item{x}{An mcmc chain.}

\item{plot_title}{The plot title.}

\item{lag_max}{The maximum lag for which to calculate auto-correlation. 
Defaults to 50.}

\item{facet}{A logical instructing the plotting of the autocorrelaiton plots
of the parameters as facetted upon parameter or else combined in one plot and
separated by colour.}
}
\value{
A plot of the auto-correlation for each parameter in x over a number 
of lag values.
}
\description{
Makes an auto-correlation plot as per the coda package, but in
the ggplot2 environment rather than base R.
}
