% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePheatmaps.R
\name{combinePheatmaps}
\alias{combinePheatmaps}
\title{Combine Pheatmaps}
\usage{
combinePheatmaps(
  ph_list,
  save_name = "joint_heatmaps.png",
  main = "heatmap_comparison",
  font_size = 20
)
}
\description{
Combine Pheatmaps
}
