% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHeatmapPSMandData.R
\name{compareHeatmapPSMandData}
\alias{compareHeatmapPSMandData}
\title{compareHeatmapPSMandData}
\usage{
compareHeatmapPSMandData(
  ...,
  save_name = "joint_heatmaps.png",
  main = "heatmap_comparison",
  col_pal_sim = (grDevices::colorRampPalette(RColorBrewer::brewer.pal(n = 7, name =
    "Blues")))(100),
  col_pal_expr = (grDevices::colorRampPalette(c("#146EB4", "white", "#FF9900")))(100),
  expr_breaks = NULL,
  sim_breaks = NULL,
  show_row_labels = FALSE,
  show_col_labels = FALSE,
  annotation_row = NA,
  annotation_colours = NA
)
}
\description{
compareHeatmapPSMandData
}
