% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHeatmaps.R
\name{compareHeatmaps}
\alias{compareHeatmaps}
\title{Compare heatmaps}
\usage{
compareHeatmaps(
  ...,
  save_name = "joint_heatmaps.png",
  main = "heatmap_comparison",
  col_pal = (grDevices::colorRampPalette(RColorBrewer::brewer.pal(n = 7, name =
    "Blues")))(100)
)
}
\description{
Compare heatmaps
}
