% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMatricesGen.R
\name{compareMatricesGen}
\alias{compareMatricesGen}
\title{Compare matrices}
\usage{
compareMatricesGen(
  ...,
  matrices = NULL,
  col_pal = NULL,
  breaks = NULL,
  order_rows = T,
  order_cols = T,
  n_row = NULL,
  n_col = NULL,
  method = "complete",
  distance = "euclidean",
  show_rownames = F,
  show_colnames = F,
  collect_legend = F,
  title = NULL
)
}
\arguments{
\item{...}{A number of matrices.}

\item{matrices}{If matrices are not given in ``...``, a list of matrices to
be heatmapped.}

\item{col_pal}{The colour palette for the heatmaps. If a single vector of 
colours, then all heatmaps share the same colour palette, but the user may 
give a vector of palettes corresponding to each matrix passed to the 
function. If ``NULL``, ``,diHelpR::defineColPal`` is called.}

\item{breaks}{Similarly to ``col_pal``, a vector of breaks for the heatmaps,
either a single vector to be used by all or else a list corresponding to the
matrices passed to the function. Defaults to 
``mdiHelpR::defineDataBreaks(m1, col_pal)`` where ``m1`` is the first matrix
passed to the function.}

\item{order_rows}{A logical instructing the function to order the rows of all
matrices based upon the ordering imposed by hclust on the first.}

\item{order_cols}{A logical instructing the function to order the columns all
matrices based upon the ordering imposed by hclust on the columns of the 
first.}

\item{method}{The linkage method used in imposing the ordering of rows and 
columns. Defaults to "complete".}

\item{distance}{The distance measure used in imposing the ordering of rows and 
columns. Defaults to "euclidean".}

\item{show_rownames}{A logical instructing inclusion of row names in the
heatmaps (default is FALSE).}

\item{show_colnames}{A logical instructing inclusion of column names in the
heatmaps (default is FALSE).}

\item{collect_legend}{Instruction to hide legend on all heatmaps and gather
(assumed) common legend to the left of the plots.}

\item{title}{The title of the final grid of plots.}

\item{col_names}{A logical instructing inclusion of column names in the
heatmaps (default is TRUE).}
}
\description{
Creates a plot with the heatmaps of m_1 and m_2 side by side.
There is an option to order the matrices based on the ordering imposed upon
m_1 by hclust.
}
