% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSimilarityMatrices.R
\name{compareSimilarityMatrices}
\alias{compareSimilarityMatrices}
\title{Compare similarity matrices}
\usage{
compareSimilarityMatrices(
  ...,
  matrices = NULL,
  col_pal = NULL,
  breaks = NULL,
  matrix_imposing_order = 1,
  order_rows = T,
  order_cols = T,
  row_order = NULL,
  col_order = NULL,
  n_row = NULL,
  n_col = NULL,
  method = "complete",
  distance = "euclidean",
  show_rownames = F,
  show_colnames = F,
  collect_legend = T,
  title = NULL,
  lb = 0,
  ub = 1,
  mid_point = 0.5 * (lb + ub)
)
}
\arguments{
\item{...}{A number of matrices.}

\item{matrices}{If matrices are not given in ``...``, a list of matrices to
be heatmapped.}

\item{col_pal}{The colour palette for the heatmaps. If a single vector of
colours, then all heatmaps share the same colour palette, but the user may
give a vector of palettes corresponding to each matrix passed to the
function. If ``NULL``, ``,diHelpR::defineColPal`` is called.}

\item{breaks}{Similarly to ``col_pal``, a vector of breaks for the heatmaps,
either a single vector to be used by all or else a list corresponding to the
matrices passed to the function. Defaults to
``mdiHelpR::defineDataBreaks(m1, col_pal)`` where ``m1`` is the first matrix
passed to the function.}

\item{matrix_imposing_order}{An integer indicating which of the matrices
passed to the function should be used in deciding upon an order.}

\item{order_rows}{A logical instructing the function to order the rows of all
matrices based upon the ordering imposed by hclust on the first.}

\item{order_cols}{A vector of indices of the order of columns in the matrices.
If ``NULL`` then column order is taken from the first matrix given.}

\item{row_order}{A vector of indices of the order of rows in the matrices.
If ``NULL`` then row order is taken from the first matrix given.}

\item{method}{The linkage method used in imposing the ordering of rows and
columns. Defaults to "complete".}

\item{distance}{The distance measure used in imposing the ordering of rows and
columns. Defaults to "euclidean".}

\item{show_rownames}{A logical instructing inclusion of row names in the
heatmaps (default is FALSE).}

\item{show_colnames}{A logical instructing inclusion of column names in the
heatmaps (default is FALSE).}

\item{collect_legend}{Instruction to hide legend on all heatmaps and gather
(assumed) common legend to the left of the plots.}

\item{title}{The title of the final grid of plots.}

\item{lb}{The lower bound used in defining breaks in each heatmap (default is
0).}

\item{ub}{The upper bound used in defining breaks in each heatmap (default is
1).}

\item{mid_point}{The mid_point used in defining breaks in each heatmap
(default is 0.5).}

\item{col_names}{A logical instructing inclusion of column names in the
heatmaps (default is TRUE).}
}
\description{
Creates a plot with the heatmaps of matrices in a grid.
There is an option to order the matrices based on the ordering imposed upon
a matrix given. It is assumed all matrices have common dimenion and that
they share the same row and column names (althoug this does not directly
affect the outcome).
}
