% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameterNames.R
\name{createParameterNames}
\alias{createParameterNames}
\title{Create parameter names}
\usage{
createParameterNames(n_datasets)
}
\arguments{
\item{n_datasets}{Integer. The number of datasets used in the analysis.}
}
\value{
A vector of strings. There will be n_datasets MassParameter_ 
variables (appended 1:n_datasets) and n_datasets choose 2 Phi_ parameters
(appended by ij, where i and j are dataset indices)
}
\description{
Creates parameter names from an MDI analysis based upon the
number of datasets used.
}
