% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataColPal.R
\name{dataColPal}
\alias{dataColPal}
\title{data colour palette}
\usage{
dataColPal(n = 100)
}
\arguments{
\item{n}{The number of shades in the palette (default is 100).}
}
\value{
A colour palette appropriate for the `colours` parameter in `pheatmap`.
}
\description{
Makes the colour palette I use for heatmaps of data.
(blue - white - orange).
}
