% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineDataBreaks.R
\name{defineDataBreaks}
\alias{defineDataBreaks}
\title{defineDataBreaks}
\usage{
defineDataBreaks(x, col_pal, mid_point = NULL)
}
\arguments{
\item{x}{The dataset of interest.}

\item{col_pal}{A colour palette (of at least 4 entries)}

\item{mid_point}{The mid point on the values for the pheatmap (defaults to 
the mean of ``min(x)`` and ``max(x)``). Useful if there is a natural boundary
(e.g. 0).}
}
\description{
Defines breaks as an input for ``pheatmap`` for a given colour 
palette (to be used as the ``color`` input in ``pheatmap``) and dataset using 
the minimum and mximum values within ``x`` to define the range for the breaks.
}
