% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNumericalFromString.R
\name{extractNumericalFromString}
\alias{extractNumericalFromString}
\title{Extract numerical from string}
\usage{
extractNumericalFromString(my_string)
}
\description{
Extract numerical from string
}
