% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findESS.R
\name{findESS}
\alias{findESS}
\title{find ESS}
\usage{
findESS(mcmc_out, test_burn_in)
}
\arguments{
\item{mcmc_out}{An MCMC object (as per the coda package)}

\item{test_burn_in}{A numrical vector of different burn in values to 
calculate the ESS at.}
}
\value{
A data.frame of the ESS for different values of burn-in.
}
\description{
Calculates the effective sample size (ESS) for an MCMC chain 
over a range of potential burn-in values.
}
