% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOrder.R
\name{findOrder}
\alias{findOrder}
\title{Find order}
\usage{
findOrder(m)
}
\arguments{
\item{m}{A matrix.}
}
\value{
A vector of integers corresponding to the row numbers of an ordered m.
}
\description{
Given a matrix, m, finds the ordering of the rows as imposed by 
hclust.
}
