% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelmanDF.R
\name{gelmanDF}
\alias{gelmanDF}
\title{Gelman data frame}
\usage{
gelmanDF(
  x,
  max_bins = 50,
  confidence = 0.95,
  transform = FALSE,
  autoburnin = TRUE
)
}
\description{
Create data suitable for a plot of the Gelman-Rubin shrinkage 
factor.
}
