% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelmanPlot.R
\name{gelmanPlot}
\alias{gelmanPlot}
\title{gelman plot}
\usage{
gelmanPlot(
  x,
  max_bins = 50,
  confidence = 0.95,
  transform = FALSE,
  autoburnin = TRUE,
  auto.layout = TRUE,
  xlab = "last iteration in chain",
  ylab = "shrink factor",
  title = "Gelman-Rubin diagnostic plot"
)
}
\description{
Create a plot of the Gelman-Rubin shrinkage factor
}
