% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGaussianDataset.R
\name{generateGaussianDataset}
\alias{generateGaussianDataset}
\title{Generate Gaussian dataset}
\usage{
generateGaussianDataset(
  cluster_means,
  std_dev,
  n,
  p,
  pi,
  row_names = paste0("Person_", 1:n),
  col_names = paste0("Gene_", 1:p)
)
}
\arguments{
\item{cluster_means}{A k-vector of cluster means defining the k clusters.}

\item{n}{The number of samples to generate in the entire dataset.}

\item{p}{The number of columns to generate in the dataset.}

\item{pi}{A k-vector of the expected proportion of points to be drawn from
each distribution.}

\item{row_names}{The row names of the generated dataset.}

\item{col_names}{The column names of the generated dataset.}
}
\description{
Generate a dataset based upon a mixture of Gaussian distributions
(with independent features).
}
