% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gewekePlot.R
\name{gewekePlot}
\alias{gewekePlot}
\title{Geweke plot}
\usage{
gewekePlot(
  x,
  frac_1 = 0.1,
  frac_2 = 0.5,
  n_bins = 20,
  p_value_threshold = 0.05,
  threshold_line_colour = "grey",
  plt_title = "Geweke diagnostic plot"
)
}
\arguments{
\item{x}{An mcmc object.}

\item{frac_1}{A fraction between 0 and 1, the fraction of samples to consider 
in the first Z-score calculation.}

\item{frac_2}{A fraction between 0 and 1, the fraction of samples from the 
end of the MCMC iterations to be considered in calculating the Z-score.}

\item{n_bins}{The number of bins of samples to be considered.}

\item{p_value_threshold}{The significance threshold included in the final 
plot.}

\item{threshold_line_colour}{The colour of the significance threshold line.}

\item{plt_title}{The title of the plot.}
}
\value{
A line graph (ggplot2 object) of the Geweke Z-score at different 
intervals.
}
\description{
Creates a plot of the Geweke Z-score compareing fractions of 
samples of the data to the final samples.
}
