\name{mdiHelpR-package}
\alias{mdiHelpR-package}
\alias{mdiHelpR}
\docType{package}
\title{\packageTitle{mdiHelpR}}
\description{\packageDescription{mdiHelpR}}
\section{Package Content}{\packageIndices{mdiHelpR}}
\author{\packageAuthor{mdiHelpR}}
\section{Maintainer}{\packageMaintainer{mdiHelpR}}
\keyword{package}
