% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{PCA plot}
\usage{
pcaPlot(x, labels = NA, n_comp = 10)
}
\arguments{
\item{x}{A matrix of the loadings of each item upon each component (the "x" 
part of the prcomp() output).}

\item{labels}{A vector of the labels associated with the rows of x.}

\item{n_comp}{The number of components to plot.}
}
\value{
A ``ggplot2'' object.
}
\description{
Plot the items loadings along the principal components.
}
