% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaSeriesPlot.R
\name{pcaSeriesPlot}
\alias{pcaSeriesPlot}
\title{PCA series plot}
\usage{
pcaSeriesPlot(
  x,
  labels,
  alpha = 0.8,
  lty = 3,
  n_comp = 10,
  include_area = F,
  ribbon_alpha = 0.2,
  ribbon_lty = 4
)
}
\arguments{
\item{x}{A matrix of the loadings of each item upon each component (the "x"
part of the prcomp() output).}

\item{labels}{A vector of the labels associated with the rows of x.}

\item{n_comp}{The number of components to plot.}

\item{include_area}{Logical indicating inclusion of [quartile_1, quartile_3]
about the median value of the clusters across components.}

\item{ribbon_alpha}{The alpha used on the ribbon geom describing the range 
about the median of the clusters.}

\item{ribbon_lty}{The linetype used on the ribbon geom describing the range 
about the median of the clusters.}
}
\value{
A ``ggplot2'' plot.
}
\description{
Plot the median loadings for each cluster along the principal
components as a series, with an option to include a ribbon describing the 
range from the first ot the third quartile around this.
}
