% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{plotSummaryStats}
\alias{plotSummaryStats}
\title{Plot Summary Stats}
\usage{
plotSummaryStats(x, main = NULL, xlab = "Mean", ylab = "Standard deviation")
}
\arguments{
\item{x}{data.frame of interest}

\item{main}{title of the plot (default is NULL).}

\item{xlab}{label for x-axis in plot (default is "Mean")}

\item{ylab}{label for y-axis in plot (defualt is "Standard deviation")}
}
\value{
A scatter plot of the Mean vs the Standard deviaiton of each feature 
of x annotated by marginal histograms. A ggExtra object.
}
\description{
Plot the mean against the standard deviation of each feature 
with marginal histograms.
}
