% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quickPSM}
\alias{quickPSM}
\title{Quick PSM}
\usage{
quickPSM(partitions, tol = 0.001)
}
\arguments{
\item{partitions}{Matrix of label assignment for data across iterations.}
}
\value{
A symmetric n x n matrix (for n rows in cluster record) describing 
the fraction of iterations for which each pairwise combination of points are
assigned the same label.
}
\description{
Constructs a similarity matrix trying assuming that there are 
a small number of modes. This function is intended to be more clever in
calculating the PSM, relying upon the poor quality exploration of many MCMC
samplers. It probably is not a good idea in low dimensions.
}
