% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdiCall.R
\name{runMDI}
\alias{runMDI}
\title{Run MDI}
\usage{
runMDI(
  datasets,
  dataTypes,
  R,
  mdipp_dir = NULL,
  K = 50,
  thin = 1,
  seed = 1,
  output = NULL,
  call = TRUE
)
}
\arguments{
\item{datasets}{A list of files as inputs to mdi++}

\item{dataTypes}{A list of data types for the dataset specific mixutre models.
Entries are paired with datasets and the orders should match.}

\item{mdipp_dir}{The directory mdi++.cpp lives in. If mdi is saved to the
PATH, then use the default of ``NULL``.}

\item{K}{The maximum number of clusters across all datasets in MDI.}

\item{thin}{The thinning factor for the MCMC samples.}

\item{seed}{The random seed controlling the algorithm.}

\item{output}{The file name to save the output too; defaults to the current
directory and a string capturing the datasets and parameters used.}

\item{call}{Logical indicating to call the generated command.}
}
\value{
A string expressing the appropriate command line call for MDI.
}
\description{
Calls mdi++.cpp (Mason et al., 2016), the C++ implementation of
Multiple Dataset Integration (Kirk et al., 2012).
}
