% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMyTheme.R
\name{setMyTheme}
\alias{setMyTheme}
\title{Set my theme}
\usage{
setMyTheme(...)
}
\arguments{
\item{...}{Arguments to pass to theme_bw().}
}
\description{
Set the theme I personally use for slides and whatnot. This is
theme_bw with the headers in a facet_wrap() set to navy with white text.
}
