% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Show}
\usage{
show(x, row_1 = 1, row_n = 4, col_1 = 1, col_n = 4)
}
\arguments{
\item{x}{data.frame of interest}

\item{row_1}{the first row shown}

\item{row_n}{the last row shown}

\item{col_1}{the first column shown}

\item{col_n}{the last column shown}
}
\value{
``x[row_1:row_n, col_1:col_n]``
}
\description{
Show the first few rows and columns of a dataset.
}
