% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simColPal.R
\name{simColPal}
\alias{simColPal}
\title{data colour palette}
\usage{
simColPal(n)
}
\arguments{
\item{n}{The number of shades in the palette (default is 100).}
}
\value{
A colour palette appropriate for the `colours` parameter in `pheatmap`.
}
\description{
Makes the colour palette I use for heatmaps of similarity 
matrices (such as posterior similarity matrices or consensus matrices). 
(white - blue).
}
