% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringMutate.R
\name{stringMutate}
\alias{stringMutate}
\title{string mutate}
\usage{
stringMutate(df, str_formula, varname)
}
\arguments{
\item{df}{A data.frame.}

\item{str_formula}{A string enclosing a formula referring to variables in df.}

\item{varname}{A string that is the name of the new variable created.}
}
\value{
A data.frame of the form of df with an additional column defined by 
varname and str_formula.
}
\description{
Given a formula captured in a string, creates a new variable 
defined by that formula. Created to allow for dynamic combinations when the 
names and number of variables present is allowed to vary.
}
