% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryDF.R
\name{summaryDF}
\alias{summaryDF}
\title{Summary data.frame}
\usage{
summaryDF(x)
}
\arguments{
\item{x}{data.frame of interest}
}
\value{
A two-column data.frame where the ith row contains the mean and 
standard deviation of the ith feature of x.
}
\description{
Create a data.frame of the means and standard deviations of each
feature.
}
