% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOrder.R
\name{findOrder}
\alias{findOrder}
\title{Find order}
\usage{
findOrder(m, method = "complete", distance = "euclidean")
}
\arguments{
\item{m}{A matrix.}

\item{method}{The linkage method used by ``hclust``.}

\item{distance}{The metric used by ``dist``.}
}
\value{
A vector of integers corresponding to the row numbers of an ordered m.
}
\description{
Given a matrix, m, finds the ordering of the rows as imposed by 
hclust.
}
